# feat: use fast-json-stringify