Aims to fix: https://github.com/fastify/aws-lambda-fastify/issues/54

After few tests i remarked a slight improvement in terms of performances. See:

Benchmarks:
```
// master
aws-lambda-fastify x 20,154 ops/sec ±9.73% (70 runs sampled)
aws-serverless-express x 2,811 ops/sec ±7.62% (68 runs sampled)
aws-serverless-fastify x 4,034 ops/sec ±3.28% (69 runs sampled)
serverless-http x 18,123 ops/sec ±5.17% (75 runs sampled)
```
```
// this branch
aws-lambda-fastify x 22,178 ops/sec ±9.67% (74 runs sampled)
aws-serverless-express x 3,061 ops/sec ±6.93% (64 runs sampled)
aws-serverless-fastify x 4,241 ops/sec ±2.49% (71 runs sampled)
serverless-http x 17,973 ops/sec ±5.43% (70 runs sampled)
```

On the other hand the event in the bench test is really small. If we use a big event like the provided v1 for example we have those results:
```
//main
aws-lambda-fastify x 14,636 ops/sec ±5.78% (72 runs sampled)
aws-serverless-express x 2,627 ops/sec ±6.22% (66 runs sampled)
aws-serverless-fastify x 3,924 ops/sec ±3.07% (69 runs sampled)
serverless-http x 15,178 ops/sec ±6.32% (70 runs sampled)
```
```
// branch
aws-lambda-fastify x 12,288 ops/sec ±7.07% (74 runs sampled)
aws-serverless-express x 2,724 ops/sec ±5.06% (65 runs sampled)
aws-serverless-fastify x 3,968 ops/sec ±2.35% (71 runs sampled)
serverless-http x 15,264 ops/sec ±5.01% (73 runs sampled)
```

So in the end i think this is not worth the shot, plus we add dependency and potential issues with the schema. Sending this PR for documentation/discussion.