@zekth 2021-06-11
> Tried your changes locally too, on my machine it was even worse for the small events....
> I just introduced a new option to prevent those headers to be added: [8a4d37f](https://github.com/fastify/aws-lambda-fastify/commit/8a4d37ff83e81d3985c40ad0cb30552e7e8eb628)
> 
> If this is ok for you, I will close this PR and the issue and create a new version?

This option works fine for me!