@adrai 2021-06-11
v1.6.0 has just been released

anyway thanks for your time and contribution