@adrai 2021-06-11
fyi: just deployed v.1.6.0 with `serializeLambdaArguments: false` to production (for a lambda function not having an api-gw-authorizer in front of it) 😉